Split btsieve::ethereum module into sub-modules
The file is getting a bit long to manage properly.
